Channels unified but dictionary problems

+++ src/main/java/it/unipi/dii/aide/mircv/utils/FileSystem.java
@@ -5,12 +5,17 @@ import java.io.*;
 import it.unipi.dii.aide.mircv.data_structures.SkipInfo;
 import org.apache.commons.io.FileUtils;
 
+import java.nio.channels.FileChannel;
 import java.util.ArrayList;
 
 import static it.unipi.dii.aide.mircv.utils.Constants.*;
 
 
 public final class FileSystem {
+
+    public static FileChannel dict_channel, docTable_channel, flags_channel, docId_channel, termFreq_channel;
+    public static FileChannel partialDict_channel, partialDocId_channel, partialTermFreq_channel;
+    public static FileChannel blocks_channel, skip_channel;
 
     private FileSystem() {
         throw new UnsupportedOperationException();
@@ -21,22 +26,60 @@ public final class FileSystem {
      * that are in resources.
      */
     public static void file_cleaner() {
-
         try {
+            // Clean or create the partial folder
             File partial_folder = new File(PARTIAL_FOLDER);
-            FileUtils.cleanDirectory(partial_folder);
-            File merged_folder = new File(MERGED_FOLDER);
-            FileUtils.cleanDirectory(merged_folder);
+            if (partial_folder.exists() && partial_folder.isDirectory()) {
+                File[] partialFiles = partial_folder.listFiles();
+                if (partialFiles != null && partialFiles.length > 0) {
+                    FileUtils.cleanDirectory(partial_folder);
+                    System.out.println("Partial folder cleaned.");
+                } else {
+                    System.out.println("Partial folder is already empty.");
+                }
+            } else {
+                if (partial_folder.mkdirs()) {
+                    System.out.println("Partial folder created.");
+                } else {
+                    System.out.println("Failed to create partial folder.");
+                    return;
+                }
+            }
 
+            // Clean or create the merged folder
+            File merged_folder = new File(MERGED_FOLDER);
+            if (merged_folder.exists() && merged_folder.isDirectory()) {
+                File[] mergedFiles = merged_folder.listFiles();
+                if (mergedFiles != null && mergedFiles.length > 0) {
+                    FileUtils.cleanDirectory(merged_folder);
+                    System.out.println("Merged folder cleaned.");
+                } else {
+                    System.out.println("Merged folder is already empty.");
+                }
+            } else {
+                if (merged_folder.mkdirs()) {
+                    System.out.println("Merged folder created.");
+                } else {
+                    System.out.println("Failed to create merged folder.");
+                    return;
+                }
+            }
+
+            // Delete FLAGS_FILE if it exists
             File flags = new File(FLAGS_FILE);
-            if(flags.exists())
+            if (flags.exists()) {
                 FileUtils.delete(flags);
+                System.out.println("FLAGS_FILE deleted.");
+            } else {
+                System.out.println("FLAGS_FILE does not exist.");
+            }
 
         } catch (IOException e) {
             e.printStackTrace();
         }
-
     }
+
+
 
     public static void delete_tempFiles() {
 
@@ -83,52 +126,52 @@ public final class FileSystem {
         return docTable.exists() && dict.exists() && docDID.exists() && docTF.exists();
     }
 
-    // function to save docids  or tf posting list into file (in order to compare before and after compression)
-    public static void saveDocsInFile(ArrayList<Integer> postings, String tempFileName) throws FileNotFoundException {
-        // Create a file
-        File outputf = new File(tempFileName);
-
-        try (PrintWriter outputWriter = new PrintWriter(outputf)) {
-            for (int i = 0; i < postings.size(); i++) {
-                printDebug("posting" + i + ": " + postings.get(i));
-                outputWriter.print(postings.get(i));
-                outputWriter.println(); // Add a newline character
-            }
-        }
-    }
-
-    public static void saveDocsInFileSkipInfo(SkipInfo si, String tempFileName) {
-        // ----------- debug file ---------------
-        File outputf = new File(tempFileName);
-
-        try(PrintWriter outputWriter = new PrintWriter(outputf);) {
-
-            outputWriter.print(si.toString());
-            outputWriter.println();
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-
-//    public static void saveStructureToFile(ArrayList<String> data, String fileName) {
-//        try (FileWriter writer = new FileWriter(DEBUG_FOLDER + fileName, false)) {
-//            for (String line : data) {
-//                writer.write(line);
-//                writer.write(System.lineSeparator());
+//    // function to save docids  or tf posting list into file (in order to compare before and after compression)
+//    public static void saveDocsInFile(ArrayList<Integer> postings, String tempFileName) throws FileNotFoundException {
+//        // Create a file
+//        File outputf = new File(tempFileName);
+//
+//        try (PrintWriter outputWriter = new PrintWriter(outputf)) {
+//            for (int i = 0; i < postings.size(); i++) {
+//                printDebug("posting" + i + ": " + postings.get(i));
+//                outputWriter.print(postings.get(i));
+//                outputWriter.println(); // Add a newline character
 //            }
-//        } catch (IOException e) {
-//            e.printStackTrace();
 //        }
 //    }
 //
-//    public static void appendStringToFile(String data, String fileName) {
-//        try (FileWriter writer = new FileWriter(DEBUG_FOLDER + fileName, true)) {
-//            writer.write(data);
-//            writer.write(System.lineSeparator());
-//        } catch (IOException e) {
-//            e.printStackTrace();
+//    public static void saveDocsInFileSkipInfo(SkipInfo si, String tempFileName) {
+//        // ----------- debug file ---------------
+//        File outputf = new File(tempFileName);
+//
+//        try(PrintWriter outputWriter = new PrintWriter(outputf);) {
+//
+//            outputWriter.print(si.toString());
+//            outputWriter.println();
+//        } catch (FileNotFoundException e) {
+//            throw new RuntimeException(e);
 //        }
 //    }
+
+
+    public static void saveStructureToFile(ArrayList<String> data, String fileName, boolean append) {
+        try (FileWriter writer = new FileWriter(DEBUG_FOLDER + fileName, append)) {
+            for (String line : data) {
+                writer.write(line);
+                writer.write(System.lineSeparator());
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public static void appendStringToFile(String data, String fileName) {
+        try (FileWriter writer = new FileWriter(DEBUG_FOLDER + fileName, true)) {
+            writer.write(data);
+            writer.write(System.lineSeparator());
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
 
 }
